Move additional tests to test_core (#559)
