+++ tests/__init__.py
@@ -20,6 +20,13 @@
     USA
 """
 
+import socket
+from functools import lru_cache
+
+
+import ifaddr
+
+
 from zeroconf.core import Zeroconf
 from zeroconf.dns import DNSIncoming
 
@@ -27,3 +34,31 @@ from zeroconf.dns import DNSIncoming
 def _inject_response(zc: Zeroconf, msg: DNSIncoming) -> None:
     """Inject a DNSIncoming response."""
     zc.handle_response(msg)
+
+
+@lru_cache(maxsize=None)
+def has_working_ipv6():
+    """Return True if if the system can bind an IPv6 address."""
+    if not socket.has_ipv6:
+        return False
+
+    try:
+        sock = socket.socket(socket.AF_INET6)
+        sock.bind(('::1', 0))
+    except Exception:
+        return False
+    finally:
+        if sock:
+            sock.close()
+
+    for iface in ifaddr.get_adapters():
+        for addr in iface.ips:
+            if addr.is_IPv6 and iface.index is not None:
+                return True
+    return False
+
+
+def _clear_cache(zc):
+    for name in zc.cache.names():
+        for record in zc.cache.entries_with_name(name):
+            zc.cache.remove(record)
